reorganized award amounts files
