+++ src/js/components/award/financialAssistance/FinancialAssistanceContent.jsx
@@ -10,7 +10,7 @@ import BaseAwardAmounts from 'models/v2/award/BaseAwardAmounts';
 import AwardHistory from 'containers/award/shared/AwardHistorySectionContainer';
 import { awardTypesWithSubawards } from 'dataMapping/award/awardHistorySection';
 import ContractGrantActivityContainer from 'containers/award/shared/ContractGrantActivityContainer';
-import AwardAmountsSection from '../shared/awardAmountsSection/AwardAmountsSection';
+import AwardAmountsSection from '../shared/awardAmounts/AwardAmountsSection';
 import AdditionalInfo from '../shared/additionalInfo/AdditionalInfo';
 import AwardOverviewLeftSection from '../shared/overview/AwardOverviewLeftSection';
 import AwardOverviewRightSection from '../shared/overview/AwardOverviewRightSection';
