+++ src/js/components/award/shared/awardAmounts/AwardAmountsTable.jsx
