+++ src/js/components/award/shared/awardAmounts/RectanglePercentVizTooltip.jsx
