+++ src/js/components/award/shared/awardAmounts/SharedBarComponents.jsx
